Modularize & doc more sync code
