+++ crates/networking/p2p/sync.rs
@@ -1,26 +1,33 @@
-mod trie_rebuild;
+mod bytecode_fetcher;
+mod state_healing;
+mod state_sync;
 mod storage_fetcher;
+mod storage_healing;
+mod trie_rebuild;
 
-
+use bytecode_fetcher::bytecode_fetcher;
 use ethrex_blockchain::error::ChainError;
 use ethrex_common::{
-    types::{AccountState, Block, BlockHash, EMPTY_KECCACK_HASH},
+    types::{Block, BlockHash},
     BigEndianHash, H256, U256, U512,
 };
-use ethrex_rlp::{decode::RLPDecode, encode::RLPEncode, error::RLPDecodeError};
+use ethrex_rlp::error::RLPDecodeError;
 use ethrex_storage::{error::StoreError, Store, STATE_TRIE_SEGMENTS};
-use ethrex_trie::{Nibbles, Node, TrieError, TrieState, EMPTY_TRIE_HASH};
-use trie_rebuild::TrieRebuilder;
-use std::{array, cmp::min, collections::BTreeMap, sync::Arc};
+use ethrex_trie::{Nibbles, Node, TrieError, TrieState};
+use state_healing::heal_state_trie;
+use state_sync::state_sync;
+use std::{array, sync::Arc};
+use storage_healing::storage_healer;
 use tokio::{
     sync::{
-        mpsc::{self, error::SendError, Receiver, Sender},
+        mpsc::{self, error::SendError},
         Mutex,
     },
-    time::{sleep, Duration, Instant},
+    time::{Duration, Instant},
 };
 use tokio_util::sync::CancellationToken;
 use tracing::{debug, info, warn};
+use trie_rebuild::TrieRebuilder;
 
 use crate::{
     kademlia::KademliaTable,
@@ -438,480 +445,6 @@ impl SyncManager {
     }
 }
 
-/// Downloads the leaf values of a Block's state trie by requesting snap state from peers
-/// Also downloads the storage tries & bytecodes for each downloaded account
-/// Receives optional checkpoints in case there was a previous snap sync process that became stale, in which
-/// case it will resume it
-/// Returns the pivot staleness status (true if stale, false if not)
-/// If the pivot is not stale by the end of the state sync then the state sync was completed succesfuly
-async fn state_sync(
-    state_root: H256,
-    store: Store,
-    peers: PeerHandler,
-    key_checkpoints: Option<[H256; STATE_TRIE_SEGMENTS]>,
-    storage_trie_rebuilder_sender: Sender<Vec<(H256, H256)>>,
-    storage_healer_sender: Sender<Vec<H256>>,
-) -> Result<bool, SyncError> {
-    // Spawn tasks to fetch each state trie segment
-    let mut state_trie_tasks = tokio::task::JoinSet::new();
-    // Spawn a task to show the state sync progress
-    let state_sync_progress = StateSyncProgress::new(Instant::now());
-    let show_progress_handle =
-        tokio::task::spawn(show_state_sync_progress(state_sync_progress.clone()));
-    for i in 0..STATE_TRIE_SEGMENTS {
-        state_trie_tasks.spawn(state_sync_segment(
-            state_root,
-            peers.clone(),
-            store.clone(),
-            i,
-            key_checkpoints.map(|chs| chs[i]),
-            state_sync_progress.clone(),
-            storage_trie_rebuilder_sender.clone(),
-            storage_healer_sender.clone(),
-        ));
-    }
-    show_progress_handle.await?;
-    // Check for pivot staleness
-    let mut stale_pivot = false;
-    let mut state_trie_checkpoint = [H256::zero(); STATE_TRIE_SEGMENTS];
-    for res in state_trie_tasks.join_all().await {
-        let (index, is_stale, last_key) = res?;
-        stale_pivot |= is_stale;
-        state_trie_checkpoint[index] = last_key;
-    }
-    // Update state trie checkpoint
-    store.set_state_trie_key_checkpoint(state_trie_checkpoint)?;
-    Ok(stale_pivot)
-}
-
-/// Downloads the leaf values of the given state trie segment by requesting snap state from peers
-/// Also downloads the storage tries & bytecodes for each downloaded account
-/// Receives an optional checkpoint from a previous state sync to resume it
-/// Returns the segment number, the pivot staleness status (true if stale, false if not), and the last downloaded key
-/// If the pivot is not stale by the end of the state sync then the state sync was completed succesfuly
-#[allow(clippy::too_many_arguments)]
-async fn state_sync_segment(
-    state_root: H256,
-    peers: PeerHandler,
-    store: Store,
-    segment_number: usize,
-    checkpoint: Option<H256>,
-    state_sync_progress: StateSyncProgress,
-    storage_trie_rebuilder_sender: Sender<Vec<(H256, H256)>>,
-    storage_healer_sender: Sender<Vec<H256>>,
-) -> Result<(usize, bool, H256), SyncError> {
-    // Resume download from checkpoint if available or start from an empty trie
-    let mut start_account_hash = checkpoint.unwrap_or(STATE_TRIE_SEGMENTS_START[segment_number]);
-    // Write initial sync progress (this task is not vital so we can detach it)
-    tokio::task::spawn(StateSyncProgress::init_segment(
-        state_sync_progress.clone(),
-        segment_number,
-        start_account_hash,
-    ));
-    // Skip state sync if we are already on healing
-    if start_account_hash == STATE_TRIE_SEGMENTS_END[segment_number] {
-        // Update sync progress (this task is not vital so we can detach it)
-        tokio::task::spawn(StateSyncProgress::end_segment(
-            state_sync_progress.clone(),
-            segment_number,
-        ));
-        return Ok((segment_number, false, start_account_hash));
-    }
-    // Spawn storage & bytecode fetchers
-    let (bytecode_sender, bytecode_receiver) = mpsc::channel::<Vec<H256>>(MAX_CHANNEL_MESSAGES);
-    let (storage_sender, storage_receiver) =
-        mpsc::channel::<Vec<(H256, H256)>>(MAX_CHANNEL_MESSAGES);
-    let bytecode_fetcher_handle = tokio::spawn(bytecode_fetcher(
-        bytecode_receiver,
-        peers.clone(),
-        store.clone(),
-    ));
-    let storage_fetcher_handle = tokio::spawn(storage_fetcher(
-        storage_receiver,
-        peers.clone(),
-        store.clone(),
-        state_root,
-        storage_trie_rebuilder_sender.clone(),
-        storage_healer_sender.clone(),
-    ));
-    info!("Starting/Resuming state trie download of segment number {segment_number} from key {start_account_hash}");
-    // Fetch Account Ranges
-    // If we reached the maximum amount of retries then it means the state we are requesting is probably old and no longer available
-    let mut stale = false;
-    loop {
-        // Update sync progress (this task is not vital so we can detach it)
-        tokio::task::spawn(StateSyncProgress::update_key(
-            state_sync_progress.clone(),
-            segment_number,
-            start_account_hash,
-        ));
-        debug!("[Segment {segment_number}]: Requesting Account Range for state root {state_root}, starting hash: {start_account_hash}");
-        if let Some((account_hashes, accounts, should_continue)) = peers
-            .request_account_range(
-                state_root,
-                start_account_hash,
-                STATE_TRIE_SEGMENTS_END[segment_number],
-            )
-            .await
-        {
-            debug!(
-                "[Segment {segment_number}]: Received {} account ranges",
-                accounts.len()
-            );
-            // Update starting hash for next batch
-            start_account_hash = *account_hashes.last().unwrap();
-            // Fetch Account Storage & Bytecode
-            let mut code_hashes = vec![];
-            let mut account_hashes_and_storage_roots = vec![];
-            for (account_hash, account) in account_hashes.iter().zip(accounts.iter()) {
-                // Build the batch of code hashes to send to the bytecode fetcher
-                // Ignore accounts without code / code we already have stored
-                if account.code_hash != *EMPTY_KECCACK_HASH
-                    && store.get_account_code(account.code_hash)?.is_none()
-                {
-                    code_hashes.push(account.code_hash)
-                }
-                // Build the batch of hashes and roots to send to the storage fetcher
-                // Ignore accounts without storage and account's which storage hasn't changed from our current stored state
-                if account.storage_root != *EMPTY_TRIE_HASH
-                    && !store.contains_storage_node(*account_hash, account.storage_root)?
-                {
-                    account_hashes_and_storage_roots.push((*account_hash, account.storage_root));
-                }
-            }
-            // Send code hash batch to the bytecode fetcher
-            if !code_hashes.is_empty() {
-                bytecode_sender.send(code_hashes).await?;
-            }
-            // Send hash and root batch to the storage fetcher
-            if !account_hashes_and_storage_roots.is_empty() {
-                storage_sender
-                    .send(account_hashes_and_storage_roots)
-                    .await?;
-            }
-            // Update Snapshot
-            store.write_snapshot_account_batch(account_hashes, accounts)?;
-            // As we are downloading the state trie in segments the `should_continue` flag will mean that there
-            // are more accounts to be fetched but these accounts may belong to the next segment
-            if !should_continue || start_account_hash >= STATE_TRIE_SEGMENTS_END[segment_number] {
-                // All accounts fetched!
-                break;
-            }
-        } else {
-            stale = true;
-            break;
-        }
-    }
-    debug!("[Segment {segment_number}]: Account Trie Fetching ended, signaling storage & bytecode fetcher process");
-    // Update sync progress (this task is not vital so we can detach it)
-    tokio::task::spawn(StateSyncProgress::end_segment(
-        state_sync_progress.clone(),
-        segment_number,
-    ));
-    // Send empty batch to signal that no more batches are incoming
-    storage_sender.send(vec![]).await?;
-    bytecode_sender.send(vec![]).await?;
-    storage_fetcher_handle.await??;
-    bytecode_fetcher_handle.await??;
-    if !stale {
-        // State sync finished before becoming stale, update checkpoint so we skip state sync on the next cycle
-        start_account_hash = STATE_TRIE_SEGMENTS_END[segment_number]
-    }
-    Ok((segment_number, stale, start_account_hash))
-}
-
-/// Waits for incoming code hashes from the receiver channel endpoint, queues them, and fetches and stores their bytecodes in batches
-async fn bytecode_fetcher(
-    mut receiver: Receiver<Vec<H256>>,
-    peers: PeerHandler,
-    store: Store,
-) -> Result<(), SyncError> {
-    let mut pending_bytecodes: Vec<H256> = vec![];
-    let mut incoming = true;
-    while incoming {
-        // Fetch incoming requests
-        match receiver.recv().await {
-            Some(code_hashes) if !code_hashes.is_empty() => {
-                pending_bytecodes.extend(code_hashes);
-            }
-            // Disconnect / Empty message signaling no more bytecodes to sync
-            _ => incoming = false,
-        }
-        // If we have enough pending bytecodes to fill a batch
-        // or if we have no more incoming batches, spawn a fetch process
-        while pending_bytecodes.len() >= BATCH_SIZE || !incoming && !pending_bytecodes.is_empty() {
-            let next_batch = pending_bytecodes
-                .drain(..BATCH_SIZE.min(pending_bytecodes.len()))
-                .collect::<Vec<_>>();
-            let remaining = fetch_bytecode_batch(next_batch, peers.clone(), store.clone()).await?;
-            // Add unfeched bytecodes back to the queue
-            pending_bytecodes.extend(remaining);
-        }
-    }
-    Ok(())
-}
-
-/// Receives a batch of code hahses, fetches their respective bytecodes via p2p and returns a list of the code hashes that couldn't be fetched in the request (if applicable)
-async fn fetch_bytecode_batch(
-    mut batch: Vec<H256>,
-    peers: PeerHandler,
-    store: Store,
-) -> Result<Vec<H256>, StoreError> {
-    if let Some(bytecodes) = peers.request_bytecodes(batch.clone()).await {
-        debug!("Received {} bytecodes", bytecodes.len());
-        // Store the bytecodes
-        for code in bytecodes.into_iter() {
-            store.add_account_code(batch.remove(0), code)?;
-        }
-    }
-    // Return remaining code hashes in the batch if we couldn't fetch all of them
-    Ok(batch)
-}
-
-/// Heals the trie given its state_root by fetching any missing nodes in it via p2p
-/// Returns true if healing was fully completed or false if we need to resume healing on the next sync cycle
-async fn heal_state_trie(
-    state_root: H256,
-    store: Store,
-    peers: PeerHandler,
-    storage_healer_sender: Sender<Vec<H256>>,
-) -> Result<bool, SyncError> {
-    let mut paths = store.get_state_heal_paths()?.unwrap_or_default();
-    // Spawn a bytecode fetcher for this block
-    let (bytecode_sender, bytecode_receiver) = mpsc::channel::<Vec<H256>>(MAX_CHANNEL_MESSAGES);
-    let bytecode_fetcher_handle = tokio::spawn(bytecode_fetcher(
-        bytecode_receiver,
-        peers.clone(),
-        store.clone(),
-    ));
-    // Add the current state trie root to the pending paths
-    paths.push(Nibbles::default());
-    while !paths.is_empty() {
-        // Spawn multiple parallel requests
-        let mut state_tasks = tokio::task::JoinSet::new();
-        for _ in 0..MAX_PARALLEL_FETCHES {
-            // Spawn fetcher for the batch
-            let batch = paths.drain(0..min(paths.len(), NODE_BATCH_SIZE)).collect();
-            state_tasks.spawn(heal_state_batch(
-                state_root,
-                batch,
-                peers.clone(),
-                store.clone(),
-                storage_healer_sender.clone(),
-                bytecode_sender.clone(),
-            ));
-            // End loop if we have no more paths to fetch
-            if paths.is_empty() {
-                break;
-            }
-        }
-        // Process the results of each batch
-        let mut stale = false;
-        for res in state_tasks.join_all().await {
-            let (return_paths, is_stale) = res?;
-            stale |= is_stale;
-            paths.extend(return_paths);
-        }
-        if stale {
-            break;
-        }
-    }
-    debug!("State Healing stopped, signaling storage healer");
-    // Save paths for the next cycle
-    if !paths.is_empty() {
-        debug!("Caching {} paths for the next cycle", paths.len());
-        store.set_state_heal_paths(paths.clone())?;
-    }
-    // Send empty batch to signal that no more batches are incoming
-    bytecode_sender.send(vec![]).await?;
-    bytecode_fetcher_handle.await??;
-    Ok(paths.is_empty())
-}
-
-/// Receives a set of state trie paths, fetches their respective nodes, stores them,
-/// and returns their children paths and the paths that couldn't be fetched so they can be returned to the queue
-/// Also returns a boolean indicating if the pivot became stale during the request
-async fn heal_state_batch(
-    state_root: H256,
-    mut batch: Vec<Nibbles>,
-    peers: PeerHandler,
-    store: Store,
-    storage_sender: Sender<Vec<H256>>,
-    bytecode_sender: Sender<Vec<H256>>,
-) -> Result<(Vec<Nibbles>, bool), SyncError> {
-    if let Some(nodes) = peers
-        .request_state_trienodes(state_root, batch.clone())
-        .await
-    {
-        debug!("Received {} state nodes", nodes.len());
-        let mut hashed_addresses = vec![];
-        let mut code_hashes = vec![];
-        // For each fetched node:
-        // - Add its children to the queue (if we don't have them already)
-        // - If it is a leaf, request its bytecode & storage
-        // - If it is a leaf, add its path & value to the trie
-        for node in nodes {
-            // We cannot keep the trie state open
-            let mut trie = store.open_state_trie(*EMPTY_TRIE_HASH);
-            let path = batch.remove(0);
-            batch.extend(node_missing_children(&node, &path, trie.state())?);
-            if let Node::Leaf(node) = &node {
-                // Fetch bytecode & storage
-                let account = AccountState::decode(&node.value)?;
-                // By now we should have the full path = account hash
-                let path = &path.concat(node.partial.clone()).to_bytes();
-                if path.len() != 32 {
-                    // Something went wrong
-                    return Err(SyncError::CorruptPath);
-                }
-                let account_hash = H256::from_slice(path);
-                if account.storage_root != *EMPTY_TRIE_HASH
-                    && !store.contains_storage_node(account_hash, account.storage_root)?
-                {
-                    hashed_addresses.push(account_hash);
-                }
-                if account.code_hash != *EMPTY_KECCACK_HASH
-                    && store.get_account_code(account.code_hash)?.is_none()
-                {
-                    code_hashes.push(account.code_hash);
-                }
-            }
-            // Add node to trie
-            let hash = node.compute_hash();
-            trie.state_mut().write_node(node, hash)?;
-        }
-        // Send storage & bytecode requests
-        if !hashed_addresses.is_empty() {
-            storage_sender.send(hashed_addresses).await?;
-        }
-        if !code_hashes.is_empty() {
-            bytecode_sender.send(code_hashes).await?;
-        }
-        Ok((batch, false))
-    } else {
-        Ok((batch, true))
-    }
-}
-
-/// Waits for incoming hashed addresses from the receiver channel endpoint and queues the associated root nodes for state retrieval
-/// Also retrieves their children nodes until we have the full storage trie stored
-/// If the state becomes stale while fetching, returns its current queued account hashes
-// Returns true if there are no more pending storages in the queue (aka storage healing was completed)
-async fn storage_healer(
-    state_root: H256,
-    mut receiver: Receiver<Vec<H256>>,
-    peers: PeerHandler,
-    store: Store,
-) -> Result<bool, SyncError> {
-    let mut pending_paths: BTreeMap<H256, Vec<Nibbles>> = store
-        .get_storage_heal_paths()?
-        .unwrap_or_default()
-        .into_iter()
-        .collect();
-    // The pivot may become stale while the fetcher is active, we will still keep the process
-    // alive until the end signal so we don't lose queued messages
-    let mut stale = false;
-    let mut incoming = true;
-    while incoming || !pending_paths.is_empty() {
-        // If we have enough pending storages to fill a batch
-        // or if we have no more incoming batches, spawn a fetch process
-        // If the pivot became stale don't process anything and just save incoming requests
-        let mut storage_tasks = tokio::task::JoinSet::new();
-        let mut task_num = 0;
-        while !stale && !pending_paths.is_empty() && task_num < MAX_PARALLEL_FETCHES {
-            let mut next_batch: BTreeMap<H256, Vec<Nibbles>> = BTreeMap::new();
-            // Fill batch
-            let mut batch_size = 0;
-            while batch_size < NODE_BATCH_SIZE && !pending_paths.is_empty() {
-                let (key, val) = pending_paths.pop_first().unwrap();
-                batch_size += val.len();
-                next_batch.insert(key, val);
-            }
-            storage_tasks.spawn(heal_storage_batch(
-                state_root,
-                next_batch.clone(),
-                peers.clone(),
-                store.clone(),
-            ));
-            task_num += 1;
-        }
-        // Add unfetched paths to queue and handle stale signal
-        for res in storage_tasks.join_all().await {
-            let (remaining, is_stale) = res?;
-            pending_paths.extend(remaining);
-            stale |= is_stale;
-        }
-
-        // Read incoming requests that are already awaiting on the receiver
-        // Don't wait for requests unless we have no pending paths left
-        if incoming && (!receiver.is_empty() || pending_paths.is_empty()) {
-            // Fetch incoming requests
-            let mut msg_buffer = vec![];
-            if receiver.recv_many(&mut msg_buffer, MAX_CHANNEL_READS).await != 0 {
-                for account_hashes in msg_buffer {
-                    if !account_hashes.is_empty() {
-                        pending_paths.extend(
-                            account_hashes
-                                .into_iter()
-                                .map(|acc_path| (acc_path, vec![Nibbles::default()])),
-                        );
-                    } else {
-                        // Empty message signaling no more bytecodes to sync
-                        incoming = false
-                    }
-                }
-            } else {
-                // Disconnect
-                incoming = false
-            }
-        }
-    }
-    let healing_complete = pending_paths.is_empty();
-    // Store pending paths
-    store.set_storage_heal_paths(pending_paths.into_iter().collect())?;
-    Ok(healing_complete)
-}
-
-/// Receives a set of storage trie paths (grouped by their corresponding account's state trie path),
-/// fetches their respective nodes, stores them, and returns their children paths and the paths that couldn't be fetched so they can be returned to the queue
-/// Also returns a boolean indicating if the pivot became stale during the request
-async fn heal_storage_batch(
-    state_root: H256,
-    mut batch: BTreeMap<H256, Vec<Nibbles>>,
-    peers: PeerHandler,
-    store: Store,
-) -> Result<(BTreeMap<H256, Vec<Nibbles>>, bool), SyncError> {
-    if let Some(mut nodes) = peers
-        .request_storage_trienodes(state_root, batch.clone())
-        .await
-    {
-        debug!("Received {} storage nodes", nodes.len());
-        // Process the nodes for each account path
-        for (acc_path, paths) in batch.iter_mut() {
-            let mut trie = store.open_storage_trie(*acc_path, *EMPTY_TRIE_HASH);
-            // Get the corresponding nodes
-            for node in nodes.drain(..paths.len().min(nodes.len())) {
-                let path = paths.remove(0);
-                // Add children to batch
-                let children = node_missing_children(&node, &path, trie.state())?;
-                paths.extend(children);
-                let hash = node.compute_hash();
-                trie.state_mut().write_node(node, hash)?;
-            }
-            // Cut the loop if we ran out of nodes
-            if nodes.is_empty() {
-                break;
-            }
-        }
-        // Return remaining and added paths to be added to the queue
-        // Filter out the storages we completely fetched
-        batch.retain(|_, v| !v.is_empty());
-        return Ok((batch, false));
-    }
-    // Pivot became stale, lets inform the fetcher
-    Ok((batch, true))
-}
-
 /// Returns the partial paths to the node's children if they are not already part of the trie state
 fn node_missing_children(
     node: &Node,
@@ -935,89 +468,6 @@ fn node_missing_children(
         _ => {}
     }
     Ok(paths)
-}
-#[derive(Clone)]
-struct StateSyncProgress {
-    data: Arc<Mutex<StateSyncProgressData>>,
-}
-
-#[derive(Clone)]
-struct StateSyncProgressData {
-    cycle_start: Instant,
-    initial_keys: [H256; STATE_TRIE_SEGMENTS],
-    current_keys: [H256; STATE_TRIE_SEGMENTS],
-    ended: [bool; STATE_TRIE_SEGMENTS],
-}
-
-impl StateSyncProgress {
-    fn new(cycle_start: Instant) -> Self {
-        Self {
-            data: Arc::new(Mutex::new(StateSyncProgressData {
-                cycle_start,
-                initial_keys: Default::default(),
-                current_keys: Default::default(),
-                ended: Default::default(),
-            })),
-        }
-    }
-
-    async fn init_segment(progress: StateSyncProgress, segment_number: usize, initial_key: H256) {
-        progress.data.lock().await.initial_keys[segment_number] = initial_key;
-    }
-    async fn update_key(progress: StateSyncProgress, segment_number: usize, current_key: H256) {
-        progress.data.lock().await.current_keys[segment_number] = current_key
-    }
-    async fn end_segment(progress: StateSyncProgress, segment_number: usize) {
-        progress.data.lock().await.ended[segment_number] = true
-    }
-
-    // Returns true if the state sync ended
-    async fn show_progress(&self) -> bool {
-        // Copy the current data so we don't read while it is being written
-        let data = self.data.lock().await.clone();
-        // Calculate the total amount of accounts synced
-        let mut synced_accounts = U256::zero();
-        // Calculate the total amount of accounts synced this cycle
-        let mut synced_accounts_this_cycle = U256::one();
-        for i in 0..STATE_TRIE_SEGMENTS {
-            let segment_synced_accounts =
-                data.current_keys[i].into_uint() - STATE_TRIE_SEGMENTS_START[i].into_uint();
-            let segment_completion_rate = (U512::from(segment_synced_accounts + 1) * 100)
-                / U512::from(U256::MAX / STATE_TRIE_SEGMENTS);
-            debug!("Segment {i} completion rate: {segment_completion_rate}%");
-            synced_accounts += segment_synced_accounts;
-            synced_accounts_this_cycle +=
-                data.current_keys[i].into_uint() - data.initial_keys[i].into_uint();
-        }
-        // Calculate current progress percentage
-        let completion_rate: U512 = (U512::from(synced_accounts) * 100) / U512::from(U256::MAX);
-        // Make a simple time to finish estimation based on current progress
-        // The estimation relies on account hashes being (close to) evenly distributed
-        let remaining_accounts =
-            (U512::from(U256::MAX) / 100) * (U512::from(100) - completion_rate);
-        // Time to finish = Time since start / Accounts synced this cycle * Remaining accounts
-        let time_to_finish_secs =
-            U512::from(Instant::now().duration_since(data.cycle_start).as_secs())
-                * remaining_accounts
-                / U512::from(synced_accounts_this_cycle);
-        info!(
-            "Downloading state trie, completion rate: {}%, estimated time to finish: {}",
-            completion_rate,
-            seconds_to_readable(time_to_finish_secs)
-        );
-        data.ended.iter().all(|e| *e)
-    }
-}
-
-async fn show_state_sync_progress(progress: StateSyncProgress) {
-    // Rest for one interval so we don't start computing on empty progress
-    sleep(SHOW_PROGRESS_INTERVAL_DURATION).await;
-    let mut interval = tokio::time::interval(SHOW_PROGRESS_INTERVAL_DURATION);
-    let mut complete = false;
-    while !complete {
-        interval.tick().await;
-        complete = progress.show_progress().await
-    }
 }
 
 fn seconds_to_readable(seconds: U512) -> String {
